requirements.txt
fastapi
uvicorn
sqlmodel
pydantic
python-dotenv
plaid-python
python-multipart
requests
psycopg2-binary
